Edit attributes to make gridview more symmetrical
