+++ app/src/main/java/appricottsoftware/clarity/sync/ClarityApp.java
@@ -3,6 +3,7 @@ package appricottsoftware.clarity.sync;
 import android.app.Application;
 import android.content.Context;
 
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
 import com.google.gson.Gson;
 
 // Keep a state for the entire app for API calls
@@ -10,6 +11,7 @@ public class ClarityApp extends Application {
 
     private static Context context;
     private static ClarityClient clarityClient;
+    private static GoogleSignInClient googleSignInClient;
     private static Gson gson;
 
     @Override
@@ -32,5 +34,17 @@ public class ClarityApp extends Application {
             gson = new Gson();
         }
         return gson;
+    }
+
+    public void setGoogleSignInClient(GoogleSignInClient googleSignInClient) {
+        this.googleSignInClient = googleSignInClient;
+    }
+
+    public GoogleSignInClient getGoogleSignInClient() {
+        return this.googleSignInClient;
+    }
+
+    public void clearGoogleSignInClient() {
+        googleSignInClient = null;
     }
 }
