+++ app/src/main/res/layout/fragment_browse.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@android:color/holo_orange_light">
+    android:background="@android:color/background_light">
 
     <TextView
         android:id="@+id/tvBrowseHeading"
@@ -17,12 +17,13 @@
         android:id="@+id/gv_browse"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginTop="16dp"
+        android:layout_marginTop="32dp"
+        android:layout_marginBottom="64dp"
         android:layout_marginStart="8dp"
         android:layout_marginEnd="8dp"
         android:columnWidth="200dp"
         android:gravity="center"
-        android:verticalSpacing="10dp"
+        android:verticalSpacing="16dp"
         android:horizontalSpacing="0dp"
         android:numColumns="2"
         android:stretchMode="columnWidth"
